+++ src/setting.rs
@@ -10,12 +10,6 @@ pub(crate) enum Setting<'src> {
   WindowsPowerShell(bool),
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize)]
-pub(crate) struct Shell<'src> {
-  pub(crate) arguments: Vec<StringLiteral<'src>>,
-  pub(crate) command: StringLiteral<'src>,
-}
-
 impl<'src> Display for Setting<'src> {
   fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
     match self {
@@ -26,17 +20,5 @@ impl<'src> Display for Setting<'src> {
       | Setting::WindowsPowerShell(value) => write!(f, "{}", value),
       Setting::Shell(shell) => write!(f, "{}", shell),
     }
-  }
-}
-
-impl<'src> Display for Shell<'src> {
-  fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
-    write!(f, "[{}", self.command)?;
-
-    for argument in &self.arguments {
-      write!(f, ", {}", argument)?;
-    }
-
-    write!(f, "]")
   }
 }
